% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_custom_tab.R
\name{ExpCustomStat}
\alias{ExpCustomStat}
\title{Customized summary statistics}
\usage{
ExpCustomStat(data,Cvar=NULL,Nvar=NULL,stat=NULL,gpby=TRUE,filt=NULL,dcast=FALSE,
value=NULL)
}
\arguments{
\item{data}{dataframe or Matrix}

\item{Cvar}{qualitative variables on which to stratify / subgroup or run categorical summaries}

\item{Nvar}{quantitative variables on which to run summary statistics for.}

\item{stat}{descriptive statistics. Sepecify which summary statistics required (Included all base stat functions like 'mean','medain','max','min','sum','IQR','sd','var',quantile like P0.1, P0.2 etc'). Also added two more stat here are 'PS' - percentage of shares and 'Prop' - column percentage}

\item{gpby}{default value is True. Group level summary will be created based on list of categorical variable. If summary required at each categorical variable level then keep this option as FALSE}

\item{filt}{filter out data while running the summary statistics. Filter can apply accross data or individual variable level using filt option. If there are multiple filters, seperate the conditons by using '^'. Ex: Nvar = c("X1","X2","X3","X4"), let say we need to exclude data X1>900 for X1 variable, X2==10 for X2 variable, Gender !='Male' for X3 variable and all data for X4 then filt should be, filt = c("X1>900"^"X2==10"^"Gender!='Male'"^all) or  c("X1>900"^"X2==10"^"Gender!='Male'"^ ^). in case if you want to keep all data for some of the variable listed in Nvar, then specify inside the filt like ^all^ or ^ ^(single space)}

\item{dcast}{fast dcast from data.table}

\item{value}{If dcast is TRUE, pass the variable name which needs to come on column}
}
\value{
summary statistics as dataframe. Usage of this function is detailed in user guide vignettes document.
}
\description{
Table of descriptive statistics. Output returns matrix object containing descriptive information on all input variables for each level or combination of levels in categorical/group variable. Also while running the analysis user can filter out the data by individual variable level or across data level.
}
\details{
Filter unique value from all the numeric variables

This will be usefull when we need to exclude some unique imputed or outliers values like '999' or '9999' or '-9' or '-1111', or '888' etc from each selected variables.

Eg:dat = data.frame(x = c(23,24,34,999,12,12,23,999,45),
                 y = c(1,3,4,999,0,999,0,8,999,0)

Exclude 999:
x = c(23,24,34,12,12,23,45)
y = c(1,3,4,0,0,8,0)

The complete functionality of `ExpCustomStat` function is detailed in vignette help page with example code.
}
\examples{
ExpCustomStat(mtcars,Cvar=c("vs","am","gear"),Nvar=c("disp","mpg"),
stat=c("Count","sum","PS"),gpby=TRUE,filt =NULL)
}
