% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_exp_numeric.R
\name{ExpKurtosis}
\alias{ExpKurtosis}
\title{Measures of Shape - Kurtosis}
\usage{
ExpKurtosis(x,type)
}
\arguments{
\item{x}{A numeric object or data.frame}

\item{type}{a character which specifies the method of computation. Options are "moment" or "excess"}
}
\value{
ExpKurtosis returns Kurtosis values
}
\description{
Measures of shape to give a detailed evaluation of data. Explains the amount and direction of skew. Kurotsis explains how tall and sharp the central peak is. Skewness has no units: but a number, like a z score
}
\examples{
ExpKurtosis(mtcars$hp,type="excess")
ExpKurtosis(mtcars$carb,type="moment")
ExpKurtosis(mtcars,type="excess")
}
\author{
dubrangala
}
