% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_exp_QQPlot.R
\name{ExpOutQQ}
\alias{ExpOutQQ}
\title{Quantile Quantile Plots}
\usage{
ExpOutQQ(data, nlim = 3, fname = NULL, Page = NULL, sample = NULL)
}
\arguments{
\item{data}{Input dataframe or data.table}

\item{nlim}{numeric variable limit}

\item{fname}{output file name. Output will be generated in PDF format}

\item{Page}{output pattern. if Page=c(3,2), It will generate 6 plots with 3 rows and 2 columns}

\item{sample}{random number of plots}
}
\value{
Normal quantile quantile plot
}
\description{
This function automatically scans through each variable and creates normal QQ plot also adds a line to a normal quantile quantile plot.
}
\examples{
CData = ISLR::Carseats
ExpOutQQ(CData,nlim=10,fname=NULL,Page=c(2,2),sample=4)
}
\seealso{
\code{\link[ggplot2:geom_qq]{geom_qq}}
}
