% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_exp_numeric.R
\name{ExpSkew}
\alias{ExpSkew}
\title{Measures of Shape - Skewness}
\usage{
ExpSkew(x,type)
}
\arguments{
\item{x}{A numeric object or data.frame}

\item{type}{a character which specifies the method of computation. Options are "moment" or "sample"}
}
\value{
ExpSkew returns Skewness values
}
\description{
Measures of shape to give a detailed evaluation of data. Explains the amount and direction of skew. Kurotsis explains how tall and sharp the central peak is. Skewness has no units: but a number, like a z score
}
\examples{
ExpSkew(mtcars,type="moment")
ExpSkew(mtcars,type="sample")
}
\author{
dubrangala
}
