% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_exp_categorical.R
\name{ExpWoeTable}
\alias{ExpWoeTable}
\title{Function provides summary statistics with weight of evidence}
\usage{
ExpWoeTable(X, Y, valueOfGood = NULL,print=FALSE,Round=2)
}
\arguments{
\item{X}{Independent categorical variable.}

\item{Y}{Binary response variable, it can take values of either 1 or 0.}

\item{valueOfGood}{Value of Y that is used as reference category.}

\item{print}{print results}

\item{Round}{rounds the values}
}
\value{
Weight of evidance summary table
}
\description{
Weight of evidence for categorical(X-independent) variable against Target variable (Y)
}
\details{
The weight of evidence tells the predictive power of an independent variable in relation to the dependent variable
}
\examples{
X = mtcars$gear
Y = mtcars$am
WOE = ExpWoeTable(X,Y,valueOfGood = 1)
}
\seealso{
\code{\link[InformationValue:WOETable]{WOETable}}
}
