% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_Overview_data.R
\name{ExpData}
\alias{ExpData}
\title{Function to generate the overview of a data frame}
\usage{
ExpData(data, type = 1)
}
\arguments{
\item{data}{a data frame}

\item{type}{Type 1 is overview of the data; Type 2 is structure of the data}
}
\description{
This function used to produce summaries of data structure and overview of the data frame.
}
\details{
This function provides overview and structure of the data frames.
\itemize{
\item Type = 1, overeview of the data (column names are "Descriptions", "Obs")
\item Type = 2, structure of the data (column names are "S.no", "Variable Name", "Variable Type", "% of Missing", "No. of Unique values")
}
}
\examples{
# Overview of the data
ExpData(data=mtcars,type=1)
# Structure of the data
ExpData(data=mtcars,type=2)
}
